// src/app/api/models/uebung-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LocalTime } from '../models/local-time';
export interface UebungDto {
  abtretenLocation?: string;
  abtretenTime?: LocalTime;
  antretenLocation?: string;
  antretenTime?: LocalTime;
  date?: string;
  id?: number;
  mitnehmen?: string;
  motto?: string;
  status?: string;
  stufeSlug?: string;
  tenue?: string;
  weiteres?: string;
}